feat: Add cancel friend request functionality
Implement cancel friend request functionality in the RecommendPeoples component. This allows users to cancel a friend request that they have sent. The cancelFriendRequest function is added to the profile actions file, and the handleCancelRequest function is updated in the RecommendPeoples component to call this function when the cancel button is clicked. This improves the user experience by providing the ability to manage friend requests more easily.