feat: Add groupId to RequestJoinGroup component
This commit adds the groupId prop to the RequestJoinGroup component in order to pass the group ID to the acceptRequestToJoinGroup function. This allows the user to accept a request to join a specific group. The handleAccept function has been updated to make use of the groupId prop when calling the acceptRequestToJoinGroup function. Additionally, a toast message is displayed to provide feedback to the user after accepting the request. The router is also refreshed to update the page after accepting the request.