feat: Add "No conversations found" message when there are no conversations
This commit adds a new message "No conversations found" to be displayed when there are no conversations in the conversations list. It improves the user experience by providing clear feedback to the user when there is no content to display. This change aligns with recent repository commits that focus on enhancing the user experience of the conversation-related pages.