chore: Update npm dependencies
This commit updates the npm dependencies in the package.json file. It adds the "@types/lodash" package with version "^4.17.5" and updates the "lodash" package to version "^4.17.21". This update ensures that the project is using the latest stable versions of these dependencies and aligns with recent repository commits that focus on updating dependencies.